+++ serde_arrow/src/test_with_arrow/items_wrapper.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use crate::{
     self as serde_arrow,
-    _impl::arrow::{_raw::schema::Schema, array::RecordBatch, datatypes::Field},
+    _impl::arrow::{_raw::schema::Schema, array::RecordBatch, datatypes::FieldRef},
     internal::error::PanicOnError,
     schema::{SchemaLike, TracingOptions},
     utils::{Item, Items},
@@ -10,14 +10,11 @@ use crate::{
 
 #[test]
 fn example() -> PanicOnError<()> {
-    use serde_arrow::schema::SerdeArrowSchema;
-
     let items: Vec<u64> = vec![1, 2, 3, 4, 5];
 
-    let fields_from_type: Vec<Field> =
-        SerdeArrowSchema::from_type::<Item<u64>>(TracingOptions::default())?.try_into()?;
-    let fields_from_samples: Vec<Field> =
-        SerdeArrowSchema::from_samples(&Items(&items), TracingOptions::default())?.try_into()?;
+    let fields_from_type = Vec::<FieldRef>::from_type::<Item<u64>>(TracingOptions::default())?;
+    let fields_from_samples =
+        Vec::<FieldRef>::from_samples(&Items(&items), TracingOptions::default())?;
 
     assert_eq!(fields_from_type, fields_from_samples);
     let fields = fields_from_type;